Create PROGRAM 26
C Program to accept the weight of a parcel in kilograms and calculate the rate per kilogram based on the following criteria:

Weight in kilograms	Rate per kilogram
First 5 kilograms	Rs. 800
Next 5 kilograms	Rs. 700
Above 10 kilograms	Rs. 500
Also input the type of the courier (‘I’ for International and ‘D’ for Domestic). If the type of the courier is International, an additional amount of Rs. 1500 is levied.